1) database    no longer has main    uses pvdbcr from pvDatabase 2) support    now ony supportMain 3) exampleLink major changes 4) several other changes
