+++ database/src/Makefile
@@ -8,7 +8,6 @@ EPICS_BASE_PVA_CORE_LIBS = pvDatabase pvaClient pvAccess pvAccessCA nt pvData ca
 
 INC += pv/exampleHelloRecord.h
 INC += pv/exampleHelloRPCRecord.h
-INC += pv/exampleSoftRecord.h
 INC += pv/exampleDatabase.h
 
 DBD += exampleDatabaseRegister.dbd
@@ -16,7 +15,6 @@ DBD += exampleSoftRecordRegister.dbd
 
 LIBRARY = exampleDatabase
 LIBSRCS += exampleHelloRecord.cpp
-LIBSRCS += exampleSoftRecord.cpp
 LIBSRCS += exampleHelloRPCRecord.cpp
 LIBSRCS += exampleDatabase.cpp
 LIBSRCS += exampleDatabaseRegister.cpp
@@ -26,11 +24,6 @@ exampleDatabase_LIBS += $(EPICS_BASE_PVA_CORE_LIBS)
 # shared library ABI version.
 SHRLIB_VERSION ?= 4.3.0
 
-
-PROD_HOST += exampleDatabaseMain
-exampleDatabaseMain_SRCS += exampleDatabaseMain.cpp
-exampleDatabaseMain_LIBS += exampleDatabase
-exampleDatabaseMain_LIBS += $(EPICS_BASE_PVA_CORE_LIBS)
 
 PROD_HOST += exampleHelloRPCClient
 exampleHelloRPCClient_SRCS += exampleHelloRPCClient.cpp
