+++ database/src/exampleSoftRecordRegister.cpp
@@ -7,24 +7,93 @@
  * @author mrk
  * @date 2013.07.24
  */
-
-
-/* Author: Marty Kraimer */
-
 #include <iocsh.h>
+#include <pv/standardField.h>
+#include <pv/standardPVField.h>
+#include <pv/timeStamp.h>
+#include <pv/pvTimeStamp.h>
+#include <pv/alarm.h>
+#include <pv/pvAlarm.h>
+#include <pv/controlSupport.h>
+#include <pv/scalarAlarmSupport.h>
 #include <pv/pvDatabase.h>
-#include <pv/channelProviderLocal.h>
 
 // The following must be the last include for code database uses
 #include <epicsExport.h>
 #define epicsExportSharedSymbols
-#include "pv/exampleSoftRecord.h"
 
 using namespace epics::pvData;
-using namespace epics::pvAccess;
 using namespace epics::pvDatabase;
-using namespace epics::exampleCPP::database;
 using namespace std;
+
+namespace epics { namespace exampleCPP { namespace database {
+class ExampleSoftRecord;
+typedef std::tr1::shared_ptr<ExampleSoftRecord> ExampleSoftRecordPtr;
+
+/**
+ * @brief A PVRecord that implements a hello service accessed via a channelPutGet request.
+ *
+ */
+class epicsShareClass ExampleSoftRecord :
+    public epics::pvDatabase::PVRecord
+{
+public:
+    POINTER_DEFINITIONS(ExampleSoftRecord);
+    /**
+     * @brief Create an instance of ExampleSoftRecord.
+     *
+     * @param recordName The name of the record.
+     * @return The new instance.
+     */
+    static ExampleSoftRecordPtr create(
+        std::string const & recordName);
+    /**
+     *  @brief Implement hello semantics.
+     */
+    virtual void process();
+    virtual ~ExampleSoftRecord() {}
+    virtual bool init() {return true;}
+private:
+    ExampleSoftRecord(std::string const & recordName,
+        epics::pvData::PVStructurePtr const & pvStructure);
+
+};
+
+ExampleSoftRecordPtr ExampleSoftRecord::create(
+    string const & recordName)
+{
+    FieldCreatePtr fieldCreate = getFieldCreate();
+    PVDataCreatePtr pvDataCreate = getPVDataCreate();
+    StandardFieldPtr standardField = getStandardField();
+    StructureConstPtr  topStructure = fieldCreate->createFieldBuilder()->
+        add("value",pvDouble) ->
+        add("reset",pvBoolean) ->
+        add("alarm",standardField->alarm()) ->
+        add("timeStamp",standardField->timeStamp()) ->
+        add("display",standardField->display()) ->
+        add("control",standardField->control())->
+        createStructure();
+    PVStructurePtr pvStructure = pvDataCreate->createPVStructure(topStructure);
+    ExampleSoftRecordPtr pvRecord(
+        new ExampleSoftRecord(recordName,pvStructure));
+    pvRecord->initPVRecord();
+    return pvRecord;
+}
+
+ExampleSoftRecord::ExampleSoftRecord(
+    string const & recordName,
+    PVStructurePtr const & pvStructure)
+: PVRecord(recordName,pvStructure)
+{
+    PVFieldPtr pvField;
+}
+
+void ExampleSoftRecord::process()
+{
+
+}
+
+}}}
 
 static const iocshArg testArg0 = { "recordName", iocshArgString };
 static const iocshArg *testArgs[] = {
@@ -38,7 +107,8 @@ static void exampleSoftRecordCallFunc(const iocshArgBuf *args)
     if(!recordName) {
         throw std::runtime_error("exampleSoftRecordCreate invalid number of arguments");
     }
-    ExampleSoftRecordPtr record = ExampleSoftRecord::create(recordName);
+    epics::exampleCPP::database::ExampleSoftRecordPtr record
+        = epics::exampleCPP::database::ExampleSoftRecord::create(recordName);
     bool result = PVDatabase::getMaster()->addRecord(record);
     if(!result) cout << "recordname" << " not added" << endl;
 }
