+++ documentation/arrayPerformance.html
@@ -6,9 +6,9 @@
   <meta http-equiv="content-type" content="text/html; charset=iso-8859-1" />
   <title>EPICS exampleCPP/arrayPerformance</title>
   <link rel="stylesheet" type="text/css"
-  href="http://epics-pvdata.sourceforge.net/base.css" />
+  href="https://mrkraimer.github.io/website/css/base.css" />
   <link rel="stylesheet" type="text/css"
-  href="http://epics-pvdata.sourceforge.net/epicsv4.css" />
+  href="https://mrkraimer.github.io/website/css/epicsv4.css" />
   <style type="text/css">
 /*<![CDATA[*/
      .about { margin-left: 3em; margin-right: 3em; font-size: .83em}
@@ -18,7 +18,7 @@
 /*]]>*/</style>
 
  <!-- Script that generates the Table of Contents -->
-  <script type="text/javascript" src="http://epics-pvdata.sourceforge.net/script/tocgen.js"></script>
+  <script type="text/javascript" src="https://mrkraimer.github.io/website/css/tocgen.js"></script>
 
 </head>
