+++ documentation/exampleCPP.html
@@ -248,11 +248,11 @@ mrk&gt; pwd
 mrk&gt; ../../bin/linux-x86_64/exampleDatabase st.cmd
 </pre>
 <h3>support</h3>
-<p>
-This creates records that use the <b>support</b> provided by 
-<b>pvDatabaseCPP</b>.
-</p>
-<p>Documentation:<a href="./support.html">support</a></p>
+See:
+<a
+      href="https://mrkraimer.github.io/website/developerGuide/recordSupport/recordSupport.html">
+      developerGuide
+</a> 
 
 <h3>json</h3>
 <p>
