--- documentation/support.html
@@ -1,147 +1,0 @@
-<?xml version="1.0" encoding="iso-8859-1"?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
-      "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head>
-  <meta http-equiv="content-type" content="text/html; charset=iso-8859-1" />
-  <title>EPICS exampleCPP/support</title>
-  <link rel="stylesheet" type="text/css"
-  href="https://mrkraimer.github.io/website/css/base.css" />
-  <link rel="stylesheet" type="text/css"
-  href="https://mrkraimer.github.io/website/css/epicsv4.css" />
-  <style type="text/css">
-/*<![CDATA[*/
-     .about { margin-left: 3em; margin-right: 3em; font-size: .83em}
-     table { margin-left: auto; margin-right: auto }
-     .diagram { text-align: center; margin: 2.5em 0 }
-     body { margin-right: 10% }
-/*]]>*/</style>
-
- <!-- Script that generates the Table of Contents -->
-  <script type="text/javascript" src="https://mrkraimer.github.io/website/css/tocgen.js"></script>
-
-</head>
-
-<body>
-
-<div class="head">
-<h1>EPICS exampleCPP/support</h1>
-<h2 class="nocount">2019.09.11</h2>
-
-<h2 class="nocount">Abstract</h2>
-<p>
-This creates  records that use the <b>support</b> provided by 
-<b>pvDatabaseCPP</b>.
-</p>
-
-</div> <!-- head -->
-
-<div id="toc">
-  <h2 class="nocount">Table of Contents</h2>
-</div>
-
-<!-- Place what you would like in the Table of Contents, inside the contents div -->
-<div id="contents" class="contents">	  
-<hr />
-
-<h2>Overview</h2>
-<p>
-<b>support</b> creates records that have the following fields:
-</p>
-<pre>
-structure
-    double value  // may be any numeric type, i. e. byte,ubyte,...
-    boolean reset
-    alarm_t alarm
-        int severity
-        int status
-        string message
-    time_t timeStamp
-        long secondsPastEpoch
-        int nanoseconds
-        int userTag
-    display_t display
-        double limitLow
-        double limitHigh
-        string description
-        string format
-        string units
-    control_t control
-        double limitLow
-        double limitHigh
-        double minStep
-    scalarAlarm_t scalarAlarm
-        boolean active
-        double lowAlarmLimit
-        double lowWarningLimit
-        double highWarningLimit
-        double highAlarmLimit
-        double hysteresis
-</pre>
-<p>Each record uses <b>controlSupport</b> and <b>scalarAlarmSupport</b>
-, which are implemented by pvDatabaseCPP.
-</p>
-<h2>Running the example</h2>
-<h3>Start the IOC</h3>
-<p>
-In a window start the IOC:
-</p>
-<pre>
-mrk> pwd
-/home/epicsv4/masterCPP/exampleCPP/support/iocBoot/support
-mrk> ../../bin/linux-x86_64/support st.cmd 
-...
-iocRun: All initialization complete
-traceRecordCreate PVRtraceRecord
-removeRecordCreate PVRremoveRecord
-addRecordCreate PVRaddRecord
-supportRecordCreate PVRsupportDouble
-supportRecordCreate PVRsupportUByte pvUByte
-epics> 
-</pre>
-<h3>Monitor changes</h3>
-<p>In a window:
-</p>
-<pre>
-mrk> pvget -r "value,alarm,control,scalarAlarm" -v -m PVRsupportDouble PVRsupportUByte
-</pre>
-<p>
-You will want to look at this window as you do the following.
-</p>
-<h3>Configure control and scalarAlarm</h3>
-<p>Do the following:
-</p>
-<pre>
-mrk> pwd
-/home/epicsv4/masterCPP/exampleCPP/support/testClient/scripts
-mrk> ./controlSigned
-mrk> ./controlUnsigned
-mrk> ./scalarAlarmSigned
-mrk> ./scalarAlarmUnsigned
-</pre>
-<h3>Issue puts</h3>
-<p>Some examples are:
-</p>
-<pre>
-mrk> pvput PVRsupportDouble 1
-mrk> pvget -r "record[process=true]field(value)" PVRsupportDouble 
-</pre>
-<p>When you issue the put you will see that the value only changed to .5.
-This is because minStep is .5.
-Thus the value is only changed by .5 each time the record is processed.
-The pvget causes an extra process.
-</p>
-<p>Now enter the commands:</p>
-<pre>
-mrk> pvput -r "control" PVRsupportDouble control='{"minStep":"0"}'
-mrk> pvput PVRsupportDouble 22
-</pre>
-<p>You will notice the value only goes to 10 because <b>limitHigh</b> is
-10. Also you will see the record go into alarm.
-</p>
-<h2>MORE LATER</h2>
-<p>More details will be provided later</p>
-
-</div> <!-- class="contents" -->
-</body>
-</html>
