+++ exampleClient/src/ntMultiMonitor.cpp
@@ -97,8 +97,8 @@ public:
             }
         }
 #else
-        shared_vector<boolean> changeFlags(multiData->getChannelChangeFlags());
-        shared_vector<boolean> isConnected(multiChannel->getIsConnected());
+        shared_vector<epics::pvData::boolean> changeFlags(multiData->getChannelChangeFlags());
+        shared_vector<epics::pvData::boolean> isConnected(multiChannel->getIsConnected());
         for(size_t ind=0; ind < values.size(); ++ind)
         {
             if(changeFlags[ind]) {
