+++ exampleLink/ioc/Db/Makefile
@@ -10,7 +10,7 @@ include $(TOP)/configure/CONFIG
 #----------------------------------------------------
 # Create and install (or just install)
 # databases, templates, substitutions like this
-DB += ao.db
+DB += ai.db
 
 #----------------------------------------------------
 # If <anyname>.db template is not named <anyname>*.template add
