+++ exampleLink/ioc/Db/ai.db
@@ -1,0 +1,13 @@
+record(ai, "$(name)")
+{
+	field(PREC, "1")
+	field(EGU, "amps")
+	field(HIHI, "8")
+	field(HIGH, "6")
+	field(LOW, "4")
+	field(LOLO, "2")
+	field(HHSV, "MAJOR")
+	field(HSV, "MINOR")
+	field(LSV, "MINOR")
+	field(LLSV, "MAJOR")
+}
