+++ exampleLink/ioc/src/exampleLinkInclude.dbd
@@ -7,5 +7,4 @@ include "putLinkScalarRegister.dbd"
 include "putLinkScalarArrayRegister.dbd"
 include "getLinkScalarRegister.dbd"
 include "getLinkScalarArrayRegister.dbd"
-include "scalarRegister.dbd"
-include "scalarArrayRegister.dbd"
+include "pvdbSpecialRegister.dbd"
