+++ exampleLink/iocBoot/Makefile
@@ -3,6 +3,5 @@ include $(TOP)/configure/CONFIG
 DIRS += $(wildcard *ioc*)
 DIRS += $(wildcard as*)
 DIRS += $(wildcard example*)
-DIRS += doubleArrayExternal
 include $(CONFIG)/RULES_DIRS
