--- exampleLink/iocBoot/doubleArrayExternal/Makefile
@@ -1,5 +1,0 @@
-TOP = ../..
-include $(TOP)/configure/CONFIG
-ARCH = $(EPICS_HOST_ARCH)
-TARGETS = envPaths
-include $(TOP)/configure/RULES.ioc
