--- exampleLink/iocBoot/doubleArrayExternal/st.cmd
@@ -1,13 +1,0 @@
-< envPaths
-
-cd ${TOP}
-
-## Register all support components
-dbLoadDatabase("dbd/exampleLink.dbd")
-exampleLink_registerRecordDeviceDriver(pdbbase)
-
-## Load record instance
-dbLoadRecords("db/dbArray.db","name=doubleArrayExternal,type=DOUBLE");
-
-cd ${TOP}/iocBoot/${IOC}
-iocInit()
