--- exampleLink/src/scalarArrayRegister.dbd
@@ -1,1 +1,0 @@
-registrar("scalarArrayRegister")
