--- exampleLink/src/scalarRegister.dbd
@@ -1,1 +1,0 @@
-registrar("scalarRegister")
