+++ support/Makefile
@@ -6,11 +6,6 @@ DIRS += configure
 DIRS += src
 src_DEPEND_DIRS = configure
 
-DIRS += ioc
-ioc_DEPEND_DIRS = src
-
-DIRS += iocBoot
-
 include $(TOP)/configure/RULES_TOP
