--- support/client/scripts/clearScalarAlarm
@@ -1,32 +1,0 @@
-#!/bin/bash
-if [ $# -ne 1 ]; then
-    echo "recordName?"
-    read recordName
-else 
-    recordName=$1
-fi
-
-lowAlarmLimit=0
-lowWarningLimit=0
-highWarningLimit=0
-highAlarmLimit=0
-hysteresis=0
-
-value=" scalarAlarm="
-value+="{\"lowAlarmLimit\":\""
-value+=$lowAlarmLimit
-value+="\",\"lowWarningLimit\":\""
-value+=$lowWarningLimit
-value+="\",\"highWarningLimit\":\""
-value+=$highWarningLimit
-value+="\",\"highAlarmLimit\":\""
-value+=$highAlarmLimit
-value+="\",\"hysteresis\":\""
-value+=$hysteresis
-value+="\"}"
-
-pvput -r "record[process=false]field(scalarAlarm)" $recordName $value
-
-
-
-
