+++ support/client/scripts/configAll
@@ -4,7 +4,5 @@
 ./scalarAlarmSigned
 ./controlUnsigned
 ./scalarAlarmUnsigned
-./configProcessRecord PVRprocessRecord add PVRsupportDouble
-./configProcessRecord PVRprocessRecord add PVRsupportUByte
-./configProcessRecord PVRprocessRecord add PVRscalarDouble
-./configProcessRecord PVRprocessRecord add PVRscalarUByte
+./configProcessRecord PVRsupportProcessRecord add PVRsupportDouble
+./configProcessRecord PVRsupportProcessRecord add PVRsupportUByte
