--- support/client/scripts/exampleArrayPlugin
@@ -1,37 +1,0 @@
-#!/bin/sh
-echo "unsigned byte array"
-pvput PVRscalarArrayUByte value=[1,2,3,4,5,6,7,8,9,10]
-pvput -r "value[array=1:3]" PVRscalarArrayUByte value=[10,20,30]
-pvget -r "value[array=1:3]" PVRscalarArrayUByte
-pvget PVRscalarArrayUByte
-pvput PVRscalarArrayUByte value=[1,2,3,4,5,6,7,8,9,10]
-pvput -r "value[array=1:2:5]" PVRscalarArrayUByte value=[10,20,30]
-pvget -r "value[array=1:2:5]" PVRscalarArrayUByte
-pvget PVRscalarArrayUByte
-echo "double array"
-pvput PVRscalarArrayDouble value='[".1",".2",".3",".4",".5",".6",".7",".8",".9","1.0"]'
-pvput -r "value[array=1:3]" PVRscalarArrayDouble value=[10,20,30]
-pvget -r "value[array=1:3]" PVRscalarArrayDouble
-pvget PVRscalarArrayDouble
-pvput PVRscalarArrayDouble value='[".1",".2",".3",".4",".5",".6",".7",".8",".9","1.0"]'
-pvput -r "value[array=1:2:5]" PVRscalarArrayDouble value=[10,20,30]
-pvget -r "value[array=1:2:5]" PVRscalarArrayDouble
-pvget PVRscalarArrayDouble
-echo "string array"
-pvput PVRscalarArrayString value='["zero","one","two","three","four","five","six","seven","eight","nine"]'
-pvput -r "value[array=1:3]" PVRscalarArrayString value='["now is zero","now is one","now is two"]'
-pvget -r "value[array=1:3]" PVRscalarArrayString
-pvget PVRscalarArrayString
-pvput PVRscalarArrayString value='["zero","one","two","three","four","five","six","seven","eight","nine"]'
-pvput -r "value[array=1:2:5]" PVRscalarArrayString value='["now is zero","now is one","now is two"]'
-pvget -r "value[array=1:2:5]" PVRscalarArrayString
-pvget PVRscalarArrayString
-echo "boolean array"
-pvput PVRscalarArrayBoolean value='["true","true","true","true","true","true","true","true","true","true"]'
-pvput -r "value[array=1:3]" PVRscalarArrayBoolean value='["false","false","false"]'
-pvget -r "value[array=1:3]" PVRscalarArrayBoolean
-pvget PVRscalarArrayBoolean
-pvput PVRscalarArrayBoolean value='["true","true","true","true","true","true","true","true","true","true"]'
-pvput -r "value[array=1:2:5]" PVRscalarArrayBoolean value='["false","false","false"]'
-pvget -r "value[array=1:2:5]" PVRscalarArrayBoolean
-pvget PVRscalarArrayBoolean
