--- support/ioc/Db/ao.db
@@ -1,20 +1,0 @@
-record(ao, "$(name)")
-{
-        field(SCAN,".1 second")
-	field(PREC, "1")
-	field(EGU, "volts")
-	field(HIHI, "8")
-	field(HIGH, "6")
-	field(LOW, "-6")
-	field(LOLO, "-8")
-        field(DRVH,"10")
-        field(DRVL,"-10")
-        field(HOPR,"10")
-        field(LOPR,"-10")
-        field(HYST,".1")
-        field(OROC,".2")
-	field(HHSV, "MAJOR")
-	field(HSV, "MINOR")
-	field(LSV, "MINOR")
-	field(LLSV, "MAJOR")
-}
