--- support/ioc/src/supportInclude.dbd
@@ -1,10 +1,0 @@
-include "base.dbd"
-include "PVAClientRegister.dbd"
-include "PVAServerRegister.dbd"
-include "registerChannelProviderLocal.dbd"
-include "qsrv.dbd"
-include "pvdbSpecialRegister.dbd"
-include "supportRecordRegister.dbd"
-include "scalarArrayRecordRegister.dbd"
-include "scalarRecordRegister.dbd"
-
