--- support/ioc/src/supportMain.cpp
@@ -1,30 +1,0 @@
-// Copyright information and license terms for this software can be
-// found in the file LICENSE that is included with the distribution
-
-/* helloPutGetMain.cpp */
-
-/**
- * @author mrk
- * @date 2013.07.24
- */
-
-#include <stddef.h>
-#include <stdlib.h>
-#include <stddef.h>
-#include <string.h>
-#include <stdio.h>
-
-#include "epicsExit.h"
-#include "epicsThread.h"
-#include "iocsh.h"
-
-int main(int argc,char *argv[])
-{
-    if(argc>=2) {    
-        iocsh(argv[1]);
-        epicsThreadSleep(.2);
-    }
-    iocsh(NULL);
-    epicsExit(0);
-    return(0);
-}
