--- support/iocBoot/Makefile
@@ -1,5 +1,0 @@
-TOP = ..
-include $(TOP)/configure/CONFIG
-DIRS += support
-include $(CONFIG)/RULES_DIRS
-
