+++ support/src/Makefile
@@ -4,20 +4,10 @@ include $(TOP)/configure/CONFIG
 
 supportSRC = $(TOP)/src
 
-EPICS_BASE_PVA_CORE_LIBS = pvDatabase pvAccess pvAccessCA pvData ca Com
+EPICS_BASE_PVA_CORE_LIBS = pvDatabase pvAccess pvData  Com
 
-DBD += supportRecordRegister.dbd
-DBD += scalarArrayRecordRegister.dbd
-DBD += scalarRecordRegister.dbd
-
-LIBRARY = support
-LIBSRCS += supportRecordRegister.cpp
-LIBSRCS += scalarArrayRecordRegister.cpp
-LIBSRCS += scalarRecordRegister.cpp
-support_LIBS += $(EPICS_BASE_PVA_CORE_LIBS)
-
-# shared library ABI version.
-SHRLIB_VERSION ?= 4.3.0
-PROD_SYS_LIBS_WIN32 += ws2_32
+PROD_HOST += supportMain
+supportMain_SRCS += supportMain.cpp
+supportMain_LIBS +=  $(EPICS_BASE_PVA_CORE_LIBS)
 
 include $(TOP)/configure/RULES
