--- support/src/scalarArrayRecordRegister.dbd
@@ -1,1 +1,0 @@
-registrar("scalarArrayRecordRegister")
