--- support/src/scalarRecordRegister.dbd
@@ -1,1 +1,0 @@
-registrar("scalarRecordRegister")
