--- support/src/supportRecordRegister.dbd
@@ -1,1 +1,0 @@
-registrar("supportRecordRegister")
